challenge2/app_web/requirements.txt
Flask
Jinja2
requests
numpy
pillow
